Move multinet.auth to multinet.api.auth
